Kasattu domain-luokkien päällekkäistä toiminnallisuutta yläluokkaan
