+++ src/main/java/lukuvinkkiKirjasto/domain/Kirja.java
@@ -1,19 +1,11 @@
 package lukuvinkkiKirjasto.domain;
 
 import java.time.LocalDate;
-import java.util.Date;
 
-public class Kirja {
+public class Kirja extends Vinkki {
 
     private String ISBN;
     private String genre;
-    private String nimi;
-    private int pituus;
-    private String linkki;
-    private String tekija;
-    private int julkaisuVuosi;
-    private LocalDate paivamaara;
-    private boolean luettu;
 
     public Kirja(String id, String genre, String nimi, int pituus, String linkki, String tekija, int julkaisuVuosi, LocalDate paivamaara, boolean luettu) {
         this.ISBN = id;
@@ -25,7 +17,6 @@ public class Kirja {
         this.julkaisuVuosi = julkaisuVuosi;
         this.paivamaara = paivamaara;
         this.luettu = luettu;
-
     }
 
     public int getJulkaisuVuosi() {
@@ -34,30 +25,6 @@ public class Kirja {
 
     public void setJulkaisuVuosi(int julkaisuVuosi) {
         this.julkaisuVuosi = julkaisuVuosi;
-    }
-
-    public boolean isLuettu() {
-        return luettu;
-    }
-
-    public void setLuettu(boolean luettu) {
-        this.luettu = luettu;
-    }
-    
-    public String getLuettuString() {
-        if (this.luettu) {
-            return "Luettu " + paivamaara;
-        } else {
-            return "Lukematon";
-        }
-    }
-
-    public int getJulkaistu() {
-        return julkaisuVuosi;
-    }
-
-    public void setJulkaistu(int julkaistu) {
-        this.julkaisuVuosi = julkaistu;
     }
 
     public String getISBN() {
@@ -74,46 +41,6 @@ public class Kirja {
 
     public void setGenre(String genre) {
         this.genre = genre;
-    }
-
-    public String getNimi() {
-        return nimi;
-    }
-
-    public void setNimi(String nimi) {
-        this.nimi = nimi;
-    }
-
-    public int getPituus() {
-        return pituus;
-    }
-
-    public void setPituus(int pituus) {
-        this.pituus = pituus;
-    }
-
-    public String getLinkki() {
-        return linkki;
-    }
-
-    public void setLinkki(String linkki) {
-        this.linkki = linkki;
-    }
-
-    public String getTekija() {
-        return tekija;
-    }
-
-    public void setTekija(String tekija) {
-        this.tekija = tekija;
-    }
-
-    public LocalDate getPaivamaara() {
-        return paivamaara;
-    }
-
-    public void setPaivamaara(LocalDate paivamaara) {
-        this.paivamaara = paivamaara;
     }
 
     public String toString() {
